refactor: Thomas' PR comments
